column pos is double now

+++ Kanban/Domain/Card/Column.cs
@@ -10,7 +10,7 @@ namespace Domain
         {
         }
         
-        public Column(string id, string name, int orderNumber) => (Id, Name, OrderNumber) = (id, name, orderNumber);
+        public Column(string id, string name, double orderNumber) => (Id, Name, OrderNumber) = (id, name, orderNumber);
 
         public Column(string name, int orderNumber) : this(Guid.NewGuid().ToString(), name, orderNumber)
         {
@@ -20,6 +20,6 @@ namespace Domain
 
         public string Name { get; } = null!;
 
-        public int OrderNumber { get; }
+        public double OrderNumber { get; }
     }
 }
